+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.2.0
+* UPD: Converted to API 1.11 to support Power BI Bookmarks
+* UPD: Integrated InteractiveService to support Power BI Bookmarks
+* UPD: Implemented the multi selection with Ctrl + Click
+
 ## 1.1.0
 * UPD: transfered to API 1.10
 * UPD: added new localization string